edited HomeActivity, ProfileActivity, AddFoodActivity

+++ app/src/main/java/com/example/nutriapp/models/getInfoByID/WeightPerServing.java
@@ -1,14 +1,14 @@
 package com.example.nutriapp.models.getInfoByID;
 
 public class WeightPerServing{
-    public int amount;
+    public double amount;
     public String unit;
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
